fix(ams): fix ams view
