--- src/main/java/fr/avenirsesr/portfolio/program/application/adapter/controller/TrainingPathController.java
@@ -1,26 +1,0 @@
-package fr.avenirsesr.portfolio.program.application.adapter.controller;
-
-import fr.avenirsesr.portfolio.program.application.adapter.dto.TrainingPathDTO;
-import fr.avenirsesr.portfolio.program.application.adapter.mapper.TrainingPathMapper;
-import fr.avenirsesr.portfolio.program.domain.port.input.TrainingPathService;
-import java.util.List;
-import lombok.AllArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@Slf4j
-@AllArgsConstructor
-@RequestMapping("/me/training-paths")
-public class TrainingPathController {
-  private final TrainingPathService trainingPathService;
-
-  @GetMapping()
-  public List<TrainingPathDTO> getAllTrainingPaths() {
-    return trainingPathService.getTrainingPathsByStudent().stream()
-        .map(TrainingPathMapper::fromDomainToDto)
-        .toList();
-  }
-}
