--- src/main/java/fr/avenirsesr/portfolio/program/application/adapter/dto/TrainingPathDTO.java
@@ -1,12 +1,0 @@
-package fr.avenirsesr.portfolio.program.application.adapter.dto;
-
-import fr.avenirsesr.portfolio.common.temporal.domain.model.enums.EDurationUnit;
-import io.swagger.v3.oas.annotations.media.Schema;
-import java.util.UUID;
-
-@Schema(requiredProperties = {"id", "name", "durationUnit", "durationCount"})
-public record TrainingPathDTO(
-    UUID id,
-    String name,
-    @Schema(ref = "#/components/schemas/EDurationUnit") EDurationUnit durationUnit,
-    Integer durationCount) {}
