--- src/main/java/fr/avenirsesr/portfolio/program/domain/port/input/TrainingPathService.java
@@ -1,8 +1,0 @@
-package fr.avenirsesr.portfolio.program.domain.port.input;
-
-import fr.avenirsesr.portfolio.program.domain.model.TrainingPath;
-import java.util.List;
-
-public interface TrainingPathService {
-  List<TrainingPath> getTrainingPathsByStudent();
-}
