+++ src/main/java/fr/avenirsesr/portfolio/program/infrastructure/adapter/repository/TrainingPathDatabaseRepository.java
@@ -5,9 +5,6 @@ import fr.avenirsesr.portfolio.program.domain.model.TrainingPath;
 import fr.avenirsesr.portfolio.program.domain.port.output.repository.TrainingPathRepository;
 import fr.avenirsesr.portfolio.program.infrastructure.adapter.mapper.TrainingPathMapper;
 import fr.avenirsesr.portfolio.program.infrastructure.adapter.model.TrainingPathEntity;
-import fr.avenirsesr.portfolio.user.domain.model.Student;
-import java.util.List;
-import java.util.stream.Collectors;
 import org.springframework.stereotype.Repository;
 
 @Repository
@@ -20,16 +17,5 @@ public class TrainingPathDatabaseRepository
     super(
         jpaRepository, jpaRepository, TrainingPathMapper::fromDomain, TrainingPathMapper::toDomain);
     this.jpaRepository = jpaRepository;
-  }
-
-  @Override
-  public List<TrainingPath> findAllTrainingPathsByStudents(Student student) {
-    return jpaRepository.findAllByStudentId(student.getId()).stream()
-        .map(TrainingPathMapper::toDomainWithoutRecursion)
-        .collect(Collectors.groupingBy(TrainingPath::getId))
-        .values()
-        .stream()
-        .map(List::getFirst)
-        .toList();
   }
 }
