+++ src/main/java/fr/avenirsesr/portfolio/student/progress/application/adapter/controller/StudentProgressController.java
@@ -1,8 +1,10 @@
 package fr.avenirsesr.portfolio.student.progress.application.adapter.controller;
 
 import fr.avenirsesr.portfolio.common.data.domain.model.SortCriteria;
+import fr.avenirsesr.portfolio.student.progress.application.adapter.dto.StudentProgressDTO;
 import fr.avenirsesr.portfolio.student.progress.application.adapter.dto.StudentProgressOverviewDTO;
 import fr.avenirsesr.portfolio.student.progress.application.adapter.dto.StudentProgressViewDTO;
+import fr.avenirsesr.portfolio.student.progress.application.adapter.mapper.StudentProgressMapper;
 import fr.avenirsesr.portfolio.student.progress.application.adapter.mapper.StudentProgressOverviewMapper;
 import fr.avenirsesr.portfolio.student.progress.application.adapter.mapper.StudentProgressViewMapper;
 import fr.avenirsesr.portfolio.student.progress.domain.port.input.StudentProgressService;
@@ -20,6 +22,13 @@ import org.springframework.web.bind.annotation.RestController;
 @RequestMapping("/me/student-progress")
 public class StudentProgressController {
   private final StudentProgressService studentProgressService;
+
+  @GetMapping
+  public List<StudentProgressDTO> getAllStudentProgress() {
+    return studentProgressService.getAllCurrentStudentProgress().stream()
+        .map(StudentProgressMapper::toDto)
+        .toList();
+  }
 
   @GetMapping("/overview")
   public List<StudentProgressOverviewDTO> getStudentProgressOverview() {
