+++ src/main/java/fr/avenirsesr/portfolio/student/progress/application/adapter/dto/StudentProgressDTO.java
@@ -1,0 +1,15 @@
+package fr.avenirsesr.portfolio.student.progress.application.adapter.dto;
+
+import fr.avenirsesr.portfolio.common.temporal.domain.model.enums.EDurationUnit;
+import io.swagger.v3.oas.annotations.media.Schema;
+import java.util.UUID;
+
+@Schema(requiredProperties = {"id", "studentId", "trainingPath"})
+public record StudentProgressDTO(UUID id, UUID studentId, TrainingPathDTO trainingPath) {
+  @Schema(requiredProperties = {"id", "name", "durationUnit", "durationCount"})
+  public record TrainingPathDTO(
+      UUID id,
+      String name,
+      @Schema(ref = "#/components/schemas/EDurationUnit") EDurationUnit durationUnit,
+      Integer durationCount) {}
+}
