--- src/test/java/fr/avenirsesr/portfolio/program/application/adapter/controller/TrainingPathControllerIT.java
@@ -1,140 +1,0 @@
-package fr.avenirsesr.portfolio.program.application.adapter.controller;
-
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
-
-import fr.avenirsesr.portfolio.common.language.domain.model.enums.ELanguage;
-import fr.avenirsesr.portfolio.common.security.infrastructure.adapter.model.AvenirsSecurityHeaders;
-import fr.avenirsesr.portfolio.common.testutils.BddLogger;
-import fr.avenirsesr.portfolio.shared.infrastructure.adapter.seeder.SeederRunner;
-import org.junit.jupiter.api.BeforeAll;
-import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.http.HttpHeaders;
-import org.springframework.http.MediaType;
-import org.springframework.test.context.ActiveProfiles;
-import org.springframework.test.web.servlet.MockMvc;
-
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-@AutoConfigureMockMvc
-@ActiveProfiles("test")
-public class TrainingPathControllerIT {
-
-  private static final String BASE_PATH = "/me/training-paths";
-
-  @Autowired private MockMvc mockMvc;
-
-  @Value("${hmac.secret-key}")
-  private String secretKey;
-
-  @Value("${user.student.payload}")
-  private String studentPayload;
-
-  @Value("${user.teacher.payload}")
-  private String teacherPayload;
-
-  @Value("${user.unknown.payload}")
-  private String unknownUserPayload;
-
-  @Value("${user.student.signature}")
-  private String studentSignature;
-
-  @Value("${user.teacher.signature}")
-  private String teacherSignature;
-
-  @Value("${user.unknown.signature}")
-  private String unknownUserSignature;
-
-  private ELanguage language = ELanguage.FRENCH;
-
-  @BeforeAll
-  static void setup(@Autowired SeederRunner seederRunner) {
-    seederRunner.run();
-  }
-
-  @Test
-  void shouldReturnAllProgramProgressForStudent() throws Exception {
-    BddLogger.given("the " + BASE_PATH + " enpoint");
-    BddLogger.when("performing a GET with a student user");
-    BddLogger.then("it should return all program progresses");
-    mockMvc
-        .perform(
-            get(BASE_PATH)
-                .header(AvenirsSecurityHeaders.SIGNED_CONTEXT, studentPayload)
-                .header(AvenirsSecurityHeaders.CONTEXT_KID, secretKey)
-                .header(AvenirsSecurityHeaders.CONTEXT_SIGNATURE, studentSignature)
-                .header(HttpHeaders.ACCEPT_LANGUAGE, language.getCode())
-                .accept(MediaType.APPLICATION_JSON))
-        .andExpect(status().isOk())
-        .andExpect(content().contentType(MediaType.APPLICATION_JSON))
-        .andExpect(jsonPath("$").isArray())
-        .andExpect(jsonPath("$[0].id").value("430f3ed4-6b38-4602-8bf3-7cccb65ebeac"))
-        .andExpect(jsonPath("$[0].name").value("Program Master Institute"))
-        .andExpect(jsonPath("$[0].durationUnit").value("TRIMESTER"))
-        .andExpect(jsonPath("$[0].durationCount").value(4));
-  }
-
-  @Test
-  void shouldReturn404WhenUserIsUnknown() throws Exception {
-    BddLogger.given("the " + BASE_PATH + " enpoint");
-    BddLogger.when("performing a GET with an unknown user");
-    BddLogger.then("it should return a 404");
-    mockMvc
-        .perform(
-            get(BASE_PATH)
-                .header(AvenirsSecurityHeaders.SIGNED_CONTEXT, unknownUserPayload)
-                .header(AvenirsSecurityHeaders.CONTEXT_KID, secretKey)
-                .header(AvenirsSecurityHeaders.CONTEXT_SIGNATURE, unknownUserSignature)
-                .header(HttpHeaders.ACCEPT_LANGUAGE, language.getCode())
-                .accept(MediaType.APPLICATION_JSON))
-        .andExpect(status().isNotFound())
-        .andExpect(content().contentType(MediaType.APPLICATION_JSON))
-        .andExpect(jsonPath("$.message").value("User not found"))
-        .andExpect(jsonPath("$.code").value("USER_NOT_FOUND"));
-  }
-
-  @Test
-  void shouldReturn403WhenUserIsNotStudentForAllProgramProgress() throws Exception {
-    BddLogger.given("the " + BASE_PATH + " enpoint");
-    BddLogger.when("performing a GET with a non student user");
-    BddLogger.then("it should return a 403");
-    mockMvc
-        .perform(
-            get(BASE_PATH)
-                .header(AvenirsSecurityHeaders.SIGNED_CONTEXT, teacherPayload)
-                .header(AvenirsSecurityHeaders.CONTEXT_KID, secretKey)
-                .header(AvenirsSecurityHeaders.CONTEXT_SIGNATURE, teacherSignature)
-                .header(HttpHeaders.ACCEPT_LANGUAGE, language.getCode())
-                .accept(MediaType.APPLICATION_JSON))
-        .andExpect(status().isForbidden())
-        .andExpect(content().contentType(MediaType.APPLICATION_JSON))
-        .andExpect(jsonPath("$.message").value("User is not student"))
-        .andExpect(jsonPath("$.code").value("USER_IS_NOT_STUDENT_EXCEPTION"));
-  }
-
-  @Test
-  void shouldFallbackInDefaultLanguageWhenLanguageNotSupportedForAllProgramProgress()
-      throws Exception {
-    BddLogger.given("the " + BASE_PATH + " enpoint");
-    BddLogger.when("performing a GET with a non supported language");
-    BddLogger.then("it should fallback in default language");
-    mockMvc
-        .perform(
-            get(BASE_PATH)
-                .header(AvenirsSecurityHeaders.SIGNED_CONTEXT, studentPayload)
-                .header(AvenirsSecurityHeaders.CONTEXT_KID, secretKey)
-                .header(AvenirsSecurityHeaders.CONTEXT_SIGNATURE, studentSignature)
-                .header(HttpHeaders.ACCEPT_LANGUAGE, "invalid_language_code")
-                .accept(MediaType.APPLICATION_JSON))
-        .andExpect(status().isOk())
-        .andExpect(content().contentType(MediaType.APPLICATION_JSON))
-        .andExpect(jsonPath("$").isArray())
-        .andExpect(jsonPath("$[0].id").value("430f3ed4-6b38-4602-8bf3-7cccb65ebeac"))
-        .andExpect(jsonPath("$[0].name").value("Program Master Institute"))
-        .andExpect(jsonPath("$[0].durationUnit").value("TRIMESTER"))
-        .andExpect(jsonPath("$[0].durationCount").value(4));
-  }
-}
