+++ src/test/java/fr/avenirsesr/portfolio/student/progress/application/adapter/controller/StudentProgressControllerIT.java
@@ -23,6 +23,7 @@ import org.springframework.test.web.servlet.MockMvc;
 @ActiveProfiles("test")
 class StudentProgressControllerIT {
 
+  private static final String BASE_PATH = "/me/student-progress";
   private static final String OVERVIEW_BASE_PATH = "/me/student-progress/overview";
   private static final String VIEW_BASE_PATH = "/me/student-progress/view";
 
@@ -227,5 +228,92 @@ class StudentProgressControllerIT {
         .andExpect(jsonPath("$[0].skills[0].levelCount").value(3))
         .andExpect(jsonPath("$[0].skills[0].currentSkillLevel").exists())
         .andExpect(jsonPath("$[0].skills[0].currentSkillLevel.traceCount").value(0));
+  }
+
+  @Test
+  void shouldReturnAllProgramProgressForStudent() throws Exception {
+    BddLogger.given("the " + BASE_PATH + " enpoint");
+    BddLogger.when("performing a GET with a student user");
+    BddLogger.then("it should return all program progresses");
+    mockMvc
+        .perform(
+            get(BASE_PATH)
+                .header(AvenirsSecurityHeaders.SIGNED_CONTEXT, studentPayload)
+                .header(AvenirsSecurityHeaders.CONTEXT_KID, secretKey)
+                .header(AvenirsSecurityHeaders.CONTEXT_SIGNATURE, studentSignature)
+                .header(HttpHeaders.ACCEPT_LANGUAGE, language.getCode())
+                .accept(MediaType.APPLICATION_JSON))
+        .andExpect(status().isOk())
+        .andExpect(content().contentType(MediaType.APPLICATION_JSON))
+        .andExpect(jsonPath("$").isArray())
+        .andExpect(jsonPath("$[0].id").value("c1206b40-7a05-4488-8496-86cb2acf40ec"))
+        .andExpect(jsonPath("$[0].studentId").value("0a8700ab-90b6-4a38-8338-acbdd4fbcd3d"))
+        .andExpect(jsonPath("$[0].trainingPath.id").value("430f3ed4-6b38-4602-8bf3-7cccb65ebeac"))
+        .andExpect(jsonPath("$[0].trainingPath.name").value("Program Master Institute"))
+        .andExpect(jsonPath("$[0].trainingPath.durationUnit").value("TRIMESTER"))
+        .andExpect(jsonPath("$[0].trainingPath.durationCount").value(4));
+  }
+
+  @Test
+  void shouldReturn404WhenUserIsUnknown() throws Exception {
+    BddLogger.given("the " + BASE_PATH + " enpoint");
+    BddLogger.when("performing a GET with an unknown user");
+    BddLogger.then("it should return a 404");
+    mockMvc
+        .perform(
+            get(BASE_PATH)
+                .header(AvenirsSecurityHeaders.SIGNED_CONTEXT, unknownUserPayload)
+                .header(AvenirsSecurityHeaders.CONTEXT_KID, secretKey)
+                .header(AvenirsSecurityHeaders.CONTEXT_SIGNATURE, unknownUserSignature)
+                .header(HttpHeaders.ACCEPT_LANGUAGE, language.getCode())
+                .accept(MediaType.APPLICATION_JSON))
+        .andExpect(status().isNotFound())
+        .andExpect(content().contentType(MediaType.APPLICATION_JSON))
+        .andExpect(jsonPath("$.message").value("User not found"))
+        .andExpect(jsonPath("$.code").value("USER_NOT_FOUND"));
+  }
+
+  @Test
+  void shouldReturn403WhenUserIsNotStudentForAllProgramProgress() throws Exception {
+    BddLogger.given("the " + BASE_PATH + " enpoint");
+    BddLogger.when("performing a GET with a non student user");
+    BddLogger.then("it should return a 403");
+    mockMvc
+        .perform(
+            get(BASE_PATH)
+                .header(AvenirsSecurityHeaders.SIGNED_CONTEXT, teacherPayload)
+                .header(AvenirsSecurityHeaders.CONTEXT_KID, secretKey)
+                .header(AvenirsSecurityHeaders.CONTEXT_SIGNATURE, teacherSignature)
+                .header(HttpHeaders.ACCEPT_LANGUAGE, language.getCode())
+                .accept(MediaType.APPLICATION_JSON))
+        .andExpect(status().isForbidden())
+        .andExpect(content().contentType(MediaType.APPLICATION_JSON))
+        .andExpect(jsonPath("$.message").value("User is not student"))
+        .andExpect(jsonPath("$.code").value("USER_IS_NOT_STUDENT_EXCEPTION"));
+  }
+
+  @Test
+  void shouldFallbackInDefaultLanguageWhenLanguageNotSupportedForAllProgramProgress()
+      throws Exception {
+    BddLogger.given("the " + BASE_PATH + " enpoint");
+    BddLogger.when("performing a GET with a non supported language");
+    BddLogger.then("it should fallback in default language");
+    mockMvc
+        .perform(
+            get(BASE_PATH)
+                .header(AvenirsSecurityHeaders.SIGNED_CONTEXT, studentPayload)
+                .header(AvenirsSecurityHeaders.CONTEXT_KID, secretKey)
+                .header(AvenirsSecurityHeaders.CONTEXT_SIGNATURE, studentSignature)
+                .header(HttpHeaders.ACCEPT_LANGUAGE, "invalid_language_code")
+                .accept(MediaType.APPLICATION_JSON))
+        .andExpect(status().isOk())
+        .andExpect(content().contentType(MediaType.APPLICATION_JSON))
+        .andExpect(jsonPath("$").isArray())
+        .andExpect(jsonPath("$[0].id").value("c1206b40-7a05-4488-8496-86cb2acf40ec"))
+        .andExpect(jsonPath("$[0].studentId").value("0a8700ab-90b6-4a38-8338-acbdd4fbcd3d"))
+        .andExpect(jsonPath("$[0].trainingPath.id").value("430f3ed4-6b38-4602-8bf3-7cccb65ebeac"))
+        .andExpect(jsonPath("$[0].trainingPath.name").value("Program Master Institute"))
+        .andExpect(jsonPath("$[0].trainingPath.durationUnit").value("TRIMESTER"))
+        .andExpect(jsonPath("$[0].trainingPath.durationCount").value(4));
   }
 }
